database is ready

+++ models/index.js
@@ -10,16 +10,16 @@ import { pool } from "../db/index.js";
  */
 
 //Note: after implmenting the AUTH we might need to change the way we are getting the data e.g not to show username, password and email - will come back to this once the auth is implemented
-// Added xid here so it is returned in object 
-export async function getFavouritesByUserId(id) {
+// Added xid here so it is returned in object
+export async function getFavouritesByUserId(uid) {
   // use pool.query to open database connection
   const res = await pool.query(
     // use the inner join to select favourites places by user's id
     // used the $1 to secure the id
-    `SELECT user_favourites.id, xid, title, city, country, suburb, description, image FROM user_favourites INNER JOIN users on user_favourites.user_id=users.id WHERE user_id= $1`,
-    [id]
+    `SELECT user_favourites.uid, xid, title, city, country, suburb, description, image FROM user_favourites INNER JOIN users on user_favourites.uid=users.uid WHERE user_favourites.uid= $1`,
+    [uid]
   );
-  console.log(res.rows)
+  console.log(res.rows);
   return res.rows;
 }
 
@@ -27,9 +27,9 @@ export async function getFavouritesByUserId(id) {
 // added xid here, and favourite.xid, so the new db entry for favourites contains xid
 export async function addNewFavourite(favourite) {
   const res = await pool.query(
-    `INSERT INTO user_favourites (user_id, xid, title, city, country, suburb, description, image) VALUES ($1, $2, $3, $4, $5, $6, $7, $8) RETURNING*;`,
+    `INSERT INTO user_favourites (uid, xid, title, city, country, suburb, description, image) VALUES ($1, $2, $3, $4, $5, $6, $7, $8) RETURNING*;`,
     [
-      favourite.user_id,
+      favourite.uid,
       favourite.xid,
       favourite.title,
       favourite.city,
@@ -45,18 +45,18 @@ export async function addNewFavourite(favourite) {
 // POST add new user
 export async function addNewUsers(users) {
   const res = await pool.query(
-    `INSERT INTO users (username, email, password) VALUES ($1, $2, $3) RETURNING*;`,
-    [users.username, users.email, users.password]
+    `INSERT INTO users (username, email, password, uid) VALUES ($1, $2, $3, $4) RETURNING*;`,
+    [users.username, users.email, users.password, users.uid]
   );
   return res.rows;
 }
 
 //DELETE - location by xid
 // changed this here so that it uses location xid instead of id
-export async function deleteFavouriteById(xid) {
+export async function deleteFavouriteById(data) {
   const res = await pool.query(
-    `DELETE FROM user_favourites WHERE xid = $1 RETURNING*;`,
-    [xid]
+    `DELETE FROM user_favourites WHERE xid = $1 AND uid = $2 RETURNING*;`,
+    [data.xid, data.uid]
   );
   return res.rows;
 }
